requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==24.1
django-extensions==4.1
python-decouple==3.8
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
pillow==10.1.0
python-dateutil==2.8.2
openpyxl==3.1.5
numpy==1.26.4
pandas==2.0.3
xlrd==2.0.1
requests